+++ .gitignore
@@ -1,3 +1,5 @@
 .ipynb_checkpoints/
 .vscode
+__pycache__
+.idea
 #*.gds
