+++ templates/arc_gc.gds
